feat: implement patterns for fields
