--- src/pages/registration/registration-style.css
@@ -1,80 +1,0 @@
-.registration-form-wrapper{
-    padding:40px 10px 20px 10px;
-    background-color: var(--main-black);
-}
-
-h1 {
-    color: var(--text-light);
-    font-size: 1.5em;
-    text-align: center;
-    padding-bottom:20px;
-    border-bottom: 1px solid var(--main-light);
-}
-.registration-form {
-
-    max-width: 700px;
-    padding: 20px;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-.account-wrapper, .birthday-wrapper{
-    width:50%;
-
-    margin-bottom:10px;
-    
-}
-
-input
-{
-width:90%;
-border:none;
-margin-top:20px;
-padding: 10px 5%;
-}
-select,
-select option{
-margin-top:20px;
-width:100%;
-padding: 10px 5%;
-}
-input[type="submit"]{
-    width:200px;
-    padding:15px;
-    margin:20px auto;
-    background-color: var(--accent-blue);
-    color: var(--text-light);
-    cursor: pointer;
-}
-input:focus{
-    border:none;
-    box-shadow: inset 0px 0px 10px var(--accent-blue);;
-}
-
-.label{
-    margin-top:20px;
-    color: var(--text-light);
-}
-
-.registration-form a {
-    color: var(--text-light);
-    text-align: center;
-}
-
-.registration-form a:hover {
-    color: var(--accent-blue);
-}
-
-.addresses-wrapper{
-    margin-top:20px;
-    border-top: 1px solid var(--main-light);
-    display: flex;
-    flex-direction: row;
-    gap:30px;
-}
-.house-index-wrapper{
-   display: flex;
-   flex-direction: row;
-   gap:20px;
-}
